pipeline_optimized/requirements.txt
# ARCO Pipeline Dependencies
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0

# BigQuery (Optional)
google-cloud-bigquery>=3.35.0
google-auth>=2.0.0

# Data Processing
pandas>=2.0.0
json5>=0.9.0